pyproject.toml
[tool.poetry]
name = "ads-backend-diploma"
version = "0.1.0"
description = "Backend for ads website"
authors = ["Serzh Ivashchenko <serzh466163@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13.3"
Django = "^4.2"
djangorestframework = "^3.14"
djangorestframework-simplejwt = "^5.3"
django-cors-headers = "^4.2"
drf-yasg = "^1.21"
django-filter = "^23.3"
Pillow = "^10.1"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
gunicorn = "^21.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.5"
pytest-cov = "^4.1"
factory-boy = "^3.3"
autopep8 = "^2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
